+++ lib/amakanize.rb
@@ -1,5 +1,6 @@
 require "amakanize/author_name"
 require "amakanize/filters/base_filter"
+require "amakanize/filters/html_unescape_filter"
 require "amakanize/filters/normalization_filter"
 require "amakanize/filters/parentheses_deletion_filter"
 require "amakanize/filters/role_name_deletion_filter"