+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Remove parentheses of numericals in author name
 - Remove role name from author name
 - Strip author name
+- Unescape HTML in author name
 
 ## 0.0.1
 - 1st release :tada: